separate out migration tests
